requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
docx==0.2.4
filelock==3.8.0
huggingface-hub==0.10.0
idna==3.4
joblib==1.2.0
lxml==4.9.1
nltk==3.7
numpy==1.23.3
packaging==21.3
pdf2image==1.16.0
Pillow==9.2.0
pyparsing==3.0.9
pytesseract==0.3.10
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.1
sentencepiece==0.1.97
threadpoolctl==3.1.0
tqdm==4.64.1
typing-extensions==4.3.0
urllib3==1.26.12
